init test_DoS.cpp
