+++ src/Makefile.test.include
@@ -49,7 +49,6 @@ BITCOIN_TESTS =\
   test/bloom_tests.cpp \
   test/checkblock_tests.cpp \
   test/coins_tests.cpp \
-  test/DoS_tests.cpp \
   test/equihash_tests.cpp \
   test/getarg_tests.cpp \
   test/key_tests.cpp \
